Conflicts:
	diana-api/src/main/java/org/openflexo/fge/BackgroundStyle.java
	diana-api/src/main/java/org/openflexo/fge/Drawing.java
	diana-api/src/main/java/org/openflexo/fge/FGEModelFactory.java
	diana-api/src/main/java/org/openflexo/fge/ForegroundStyle.java
	diana-api/src/main/java/org/openflexo/fge/graphics/FGEGeometricGraphics.java
	diana-api/src/main/java/org/openflexo/fge/shapes/ComplexCurve.java
	diana-api/src/main/java/org/openflexo/fge/shapes/ShapeSpecification.java
	diana-core/src/main/java/org/openflexo/fge/FGEModelFactoryImpl.java
	diana-core/src/main/java/org/openflexo/fge/control/DianaInteractiveEditor.java
	diana-core/src/main/java/org/openflexo/fge/control/DianaToolFactory.java
	diana-core/src/main/java/org/openflexo/fge/control/tools/DianaPalette.java
	diana-core/src/main/java/org/openflexo/fge/control/tools/PaletteController.java
	diana-core/src/main/java/org/openflexo/fge/control/tools/ShapeSpecificationFactory.java
	diana-core/src/main/java/org/openflexo/fge/graphics/FGEGraphicsImpl.java
	diana-core/src/main/java/org/openflexo/fge/impl/DrawingTreeNodeImpl.java
	diana-core/src/main/java/org/openflexo/fge/impl/GeometricNodeImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/ArcImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/ChevronImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/ComplexCurveImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/OvalImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/ParallelogramImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/PlusImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/PolygonImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/RectangleImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/RectangularOctogonImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/RegularPolygonImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/ShapeSpecificationImpl.java
	diana-core/src/main/java/org/openflexo/fge/shapes/impl/StarImpl.java
	diana-drawing-editor/src/main/java/org/openflexo/fge/drawingeditor/DiagramEditorPalette.java
	diana-drawing-editor/src/main/java/org/openflexo/fge/drawingeditor/LaunchDiagramEditor.java
	diana-geom/src/main/java/org/openflexo/fge/geom/AreaComputation.java
	diana-geom/src/main/java/org/openflexo/fge/geom/FGEArc.java
	diana-geom/src/main/java/org/openflexo/fge/geom/FGEComplexCurve.java
	diana-geom/src/main/java/org/openflexo/fge/geom/FGEEllips.java
	diana-geom/src/main/java/org/openflexo/fge/geom/FGEGeneralShape.java
	diana-geom/src/main/java/org/openflexo/fge/geom/FGEPolygon.java
	diana-geom/src/main/java/org/openflexo/fge/geom/FGERectangle.java
	diana-geom/src/main/java/org/openflexo/fge/geom/FGERoundRectangle.java
	diana-geom/src/main/java/org/openflexo/fge/geom/FGEShape.java
	diana-geom/src/main/java/org/openflexo/fge/geom/area/FGESubstractionArea.java
	diana-geom/src/main/java/org/openflexo/fge/geom/area/FGEUnionArea.java
	diana-geom/src/main/java/org/openflexo/fge/graphics/AbstractFGEGraphics.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/DiagramEditorPalette.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/GeomEditApplication.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/BandWithTwoLinesConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/CircleReference.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/CircleWithCenterAndPointConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/ComplexCurveWithNPointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/CubicCurveWithFourPointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/EllipsReference.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/GeometricConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/HalfBandWithLinesConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/HalfLineWithTwoPointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/HalfPlaneWithLineAndPointConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/HorizontalLineWithPointConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/IntersectionConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/LineIntersectionPointConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/LineReference.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/LineWithTwoPointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/NearestPointFromObjectConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/ObjectReference.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/OrthogonalLineWithPointConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/ParallelLineWithPointConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/PointMiddleOfTwoPointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/PointReference.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/PolygonWithNPointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/PolylinWithNPointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/QuadCurveWithThreePointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/RectPolylinWithStartAndEndAreaConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/RectangleWithTwoPointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/RotatedLineWithPointConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/RoundRectangleWithTwoPointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/SegmentReference.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/SegmentWithTwoPointsConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/SymetricPointConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/TangentLineWithCircleAndPointConstruction.java
	diana-geomedit/src/main/java/org/openflexo/diana/geomedit/model/VerticalLineWithPointConstruction.java
	diana-swing/src/main/java/org/openflexo/fge/swing/control/SwingToolFactory.java
	diana-swing/src/main/java/org/openflexo/fge/swing/control/tools/JDianaPalette.java
	diana-swing/src/main/java/org/openflexo/fge/swing/graphics/JFGEShapeGraphics.java
	diana-swing/src/main/java/org/openflexo/fge/swing/paint/FGEPaintManager.java
	diana-swing/src/main/java/org/openflexo/fge/swing/view/JDrawingView.java
	diana-swing/src/main/java/org/openflexo/fge/swing/view/JPaletteElementView.java

+++ diana-swing/src/main/java/org/openflexo/diana/swing/view/JDrawingView.java
@@ -203,7 +203,7 @@ public class JDrawingView<M> extends JDianaLayeredView<M> implements Autoscroll,
 
 		/*if (getController() instanceof DianaInteractiveEditor) {
 			if (((DianaInteractiveEditor<?, ?, ?>) controller).getPalettes() != null) {
-				for (DrawingPalette p : ((DianaInteractiveEditor<?, ?, ?>) controller).getPalettes()) {
+				for (PaletteModel p : ((DianaInteractiveEditor<?, ?, ?>) controller).getPalettes()) {
 					activatePalette(p);
 				}
 			}
@@ -715,13 +715,15 @@ public class JDrawingView<M> extends JDianaLayeredView<M> implements Autoscroll,
 				}
 			});
 			for (GeometricNode<?> gn : geomList) {
-				// TODO: use the same graphics, just change DrawingTreeNode
-				// JFGEGeometricGraphics geometricGraphics = new JFGEGeometricGraphics(gn, this);
-				JDianaGeometricGraphics geometricGraphics = getGeometricGraphics(gn);
-				geometricGraphics.createGraphics(g2/*, controller*/);
-				gn.paint(geometricGraphics);
-				geometricGraphics.releaseGraphics();
-				geometricGraphics.delete();
+				if (gn.shouldBeDisplayed()) {
+					// TODO: use the same graphics, just change DrawingTreeNode
+					// JDianaGeometricGraphics geometricGraphics = new JDianaGeometricGraphics(gn, this);
+					JDianaGeometricGraphics geometricGraphics = getGeometricGraphics(gn);
+					geometricGraphics.createGraphics(g2/*, controller*/);
+					gn.paint(geometricGraphics);
+					geometricGraphics.releaseGraphics();
+					geometricGraphics.delete();
+				}
 			}
 		}
 	}
